feat: add create playlist and refactor code
