+++ lib/widgets/home_widgets/section_error.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class SectionError extends StatelessWidget {
+  final String message;
+  final VoidCallback? onRetry;
+
+  const SectionError({super.key, required this.message, this.onRetry});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 120,
+      child: Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(message, style: const TextStyle(color: Colors.grey)),
+            if (onRetry != null) ...[
+              const SizedBox(height: 8),
+              TextButton(onPressed: onRetry, child: const Text('Retry')),
+            ],
+          ],
+        ),
+      ),
+    );
+  }
+}
